Create _posts
Created the first blog post for the GitHub Pages site. The post includes an introduction to the blog and a brief overview of the author's coding journey. Future plans for the blog are also highlighted, inviting readers to leave comments and suggest topics for future posts. The post is formatted using Markdown and includes YAML front matter with the title and date.